requirements.txt
faster_whisper==1.1.1
ffmpeg==1.4
ffmpeg_python==0.2.0
numpy==2.3.0
openai==1.86.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pydub==0.25.1
python-dotenv==1.1.0
pytubefix==9.1.1
silero_vad==5.1.2
tiktoken==0.9.0
torch==2.5.0+cu121
torchaudio==2.5.0+cu121
torchvision
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
ultralytics==8.3.153
pydantic
--extra-index-url https://download.pytorch.org/whl/cu121